[fix]get request according to page
